sql-agent-master/pyproject.toml
[project]
name = "sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-nebius>=0.1.3",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
